+++ crates/model/src/python/events/order/canceled.rs
@@ -136,7 +136,7 @@ impl OrderCanceled {
     #[getter]
     #[pyo3(name = "reconciliation")]
     fn py_reconciliation(&self) -> bool {
-        self.reconciliation != 0
+        self.reconciliation
     }
 
     #[pyo3(name = "to_dict")]