+++ crates/model/src/python/events/order/triggered.rs
@@ -138,7 +138,7 @@ impl OrderTriggered {
     #[getter]
     #[pyo3(name = "reconciliation")]
     fn py_reconciliation(&self) -> bool {
-        self.reconciliation != 0
+        self.reconciliation
     }
 
     #[pyo3(name = "to_dict")]