+++ crates/model/src/python/events/order/updated.rs
@@ -171,7 +171,7 @@ impl OrderUpdated {
     #[getter]
     #[pyo3(name = "reconciliation")]
     fn py_reconciliation(&self) -> bool {
-        self.reconciliation != 0
+        self.reconciliation
     }
 
     #[pyo3(name = "to_dict")]